Fix for auto-repeating media
The auto-repeat functionality was simply invoking setMedia again
with the current media. This used to be the correct thing to do.

However, the media player was changed to release (free) the native
media resource when setMedia was called. So, on a repeat play,
set the new media, release the media (and the associated native
media), play the (released) media, and crash.

The solution is to factor out the minimum code needed for the
repeat play and have the reset() method invoke only this code, and
not the other code that releases the media before setting it again.