Refactor the Info class
This class previously logged out various useful pieces of diagnostic
information (system properties, environment variables and the like).

This has now been converted to a value object and is no longer
automatically dumped out - it is up to the client application how it
wants to use (or not use) this information.

Changed the test base class to dump the information out to the
console.